<FEAT> 특정 사용자의 동아리 목록 가져오기 API 개발

+++ src/main/java/com/ajouton/noname/domain/club/entity/Club.java
@@ -19,7 +19,7 @@ public class Club extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "club_id")
-    private int clubId;
+    private Long clubId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="club_category")
